Add aktualizr update control allow block mechanism
Make it possible for the user to momentarily block the update via an
advisory lock on /run/lock/aktualizr-lock (see flock man page to
understand how to apply or remove an advisory lock on an open file).

Also add a patch file for the changes done to libaktualizr needed for this.

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>
Signed-off-by: Jeremias Cordoba <jeremias.cordoba@toradex.com>