aktualizr: introduce device data proxy
The proxy will open a local TCP port and wait for JSON formatted data on
it. Only connections from localhost are accepted. Received data will be
sent to Torizon OTA via libaktualizr SendDeviceData() API.

The proxy is disabled by default, and can be enabled with
--enable-data-proxy command line parameter. The local TCP port used by
default is 8850, but can be changed via --data-proxy-port command line
parameter.

Related-to: TOR-1855

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>